+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/test/java/com/tencent/cloud/polaris/circuitbreaker/gateway/PolarisCircuitBreakerGatewayIntegrationTest.java
@@ -57,7 +57,6 @@ import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWeb
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.cloud.client.circuitbreaker.CircuitBreakerFactory;
 import org.springframework.cloud.client.circuitbreaker.Customizer;
-import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
 import org.springframework.cloud.gateway.filter.factory.SpringCloudCircuitBreakerFilterFactory;
 import org.springframework.cloud.gateway.route.RouteLocator;
 import org.springframework.cloud.gateway.route.builder.RouteLocatorBuilder;
@@ -85,12 +84,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 				"spring.cloud.gateway.enabled=true",
 				"spring.cloud.polaris.namespace=" + NAMESPACE_TEST,
 				"spring.cloud.polaris.service=test",
-				"spring.main.web-application-type=reactive",
-				"httpbin=http://localhost:${wiremock.server.port}"
+				"spring.main.web-application-type=reactive"
 		},
 		classes = PolarisCircuitBreakerGatewayIntegrationTest.TestApplication.class
 )
-@AutoConfigureWireMock(port = 0)
 @ActiveProfiles("test-gateway")
 @AutoConfigureWebTestClient(timeout = "1000000")
 public class PolarisCircuitBreakerGatewayIntegrationTest {