+++ spring-cloud-tencent-examples/quickstart-example/quickstart-caller-service/src/main/java/com/tencent/cloud/quickstart/caller/circuitbreaker/CircuitBreakerController.java
@@ -17,6 +17,7 @@
 
 package com.tencent.cloud.quickstart.caller.circuitbreaker;
 
+import com.tencent.cloud.common.metadata.MetadataContext;
 import reactor.core.publisher.Mono;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,6 +29,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.client.HttpServerErrorException;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.reactive.function.client.WebClient;
 
@@ -110,7 +113,7 @@ public class CircuitBreakerController {
 	@GetMapping("/rest")
 	public String circuitBreakRestTemplate() {
 		return circuitBreakerFactory
-				.create("QuickstartCalleeService#/quickstart/callee/circuitBreak")
+				.create(MetadataContext.LOCAL_NAMESPACE + "#QuickstartCalleeService#/quickstart/callee/circuitBreak#http#GET")
 				.run(() -> defaultRestTemplate.getForObject("/quickstart/callee/circuitBreak", String.class),
 						throwable -> "trigger the refuse for service callee."
 				);
@@ -142,7 +145,12 @@ public class CircuitBreakerController {
 	 */
 	@GetMapping("/rest/fallbackFromPolaris")
 	public ResponseEntity<String> circuitBreakRestTemplateFallbackFromPolaris() {
-		return restTemplateFallbackFromPolaris.getForEntity("/quickstart/callee/circuitBreak", String.class);
+		try {
+			return restTemplateFallbackFromPolaris.getForEntity("/quickstart/callee/circuitBreak", String.class);
+		}
+		catch (HttpClientErrorException | HttpServerErrorException httpClientErrorException) {
+			return new ResponseEntity<>(httpClientErrorException.getResponseBodyAsString(), httpClientErrorException.getStatusCode());
+		}
 	}
 
 	/**
@@ -151,7 +159,12 @@ public class CircuitBreakerController {
 	 */
 	@GetMapping("/rest/fallbackFromCode")
 	public ResponseEntity<String> circuitBreakRestTemplateFallbackFromCode() {
-		return restTemplateFallbackFromCode.getForEntity("/quickstart/callee/circuitBreak", String.class);
+		try {
+			return restTemplateFallbackFromCode.getForEntity("/quickstart/callee/circuitBreak", String.class);
+		}
+		catch (HttpClientErrorException | HttpServerErrorException httpClientErrorException) {
+			return new ResponseEntity<>(httpClientErrorException.getResponseBodyAsString(), httpClientErrorException.getStatusCode());
+		}
 	}
 
 	/**
@@ -168,7 +181,7 @@ public class CircuitBreakerController {
 				.bodyToMono(String.class)
 				.transform(it ->
 						reactiveCircuitBreakerFactory
-								.create("QuickstartCalleeService#/quickstart/callee/circuitBreak")
+								.create(MetadataContext.LOCAL_NAMESPACE + "QuickstartCalleeService#/quickstart/callee/circuitBreak#http#GET")
 								.run(it, throwable -> Mono.just("fallback: trigger the refuse for service callee"))
 				);
 	}