+++ CHANGELOG.md
@@ -29,3 +29,4 @@
 - [fix:fix contract reporting bug when using TSF.](https://github.com/Tencent/spring-cloud-tencent/pull/1406)
 - [fix: fix PolarisCircuitBreakerConfiguration not clear when gateway invoke by wildcard apis](https://github.com/Tencent/spring-cloud-tencent/pull/1425)
 - [fix:fix actuator name warning from #1428 .](https://github.com/Tencent/spring-cloud-tencent/pull/1432)
+- [feat:upgrade api circuit breaker.](https://github.com/Tencent/spring-cloud-tencent/pull/1441)